+++ apps/tergite-mss/tests/api/rest/v2/test_me.py
@@ -5,7 +5,7 @@ from datetime import datetime, timedelta, timezone
 import pytest
 from pytest_lazyfixture import lazy_fixture
 
-from services.auth import AppToken, AuthProvider
+from services.auth import AppToken, Project
 from tests._utils.auth import (
     TEST_APP_TOKEN_DICT,
     TEST_NO_QPU_APP_TOKEN_DICT,
@@ -40,6 +40,16 @@ _MY_PROJECT_REQUESTS = [
     (user_id, get_auth_cookie(user_id), project)
     for project in PROJECT_V2_LIST
     for user_id in project["user_ids"]
+]
+_MY_ADMINISTERED_PROJECT_REQUESTS = [
+    (project["admin_id"], get_auth_cookie(project["admin_id"]), project)
+    for project in PROJECT_V2_LIST
+]
+_MY_NON_ADMINISTERED_PROJECT_REQUESTS = [
+    (user_id, get_auth_cookie(user_id), project)
+    for project in PROJECT_V2_LIST
+    for user_id in project["user_ids"]
+    if user_id != project["admin_id"]
 ]
 _MY_USER_INFO_REQUESTS = [
     (user, get_auth_cookie(str(user["_id"])))
@@ -138,9 +148,9 @@ def test_view_my_project_in_less_detail(
 
 @pytest.mark.parametrize("user_id, cookies, project", _OTHERS_PROJECT_REQUESTS)
 def test_view_others_project_is_not_allowed(
-    user_id, cookies, project, client_v2, inserted_projects
+    user_id, cookies, project, client_v2, inserted_projects_v2
 ):
-    """No user can view only other's projects at /v2/me/projects/{id}"""
+    """No user can view other's projects at /v2/me/projects/{id}"""
     # using context manager to ensure on_startup runs
     with client_v2 as client:
         project_id = project["_id"]
@@ -148,10 +158,50 @@ def test_view_others_project_is_not_allowed(
         response = client.get(url, cookies=cookies)
 
         got = response.json()
-        expected = {"detail": "Not Found"}
+        expected = {"detail": "the project does not exist."}
 
         assert response.status_code == 404
         assert got == expected
+
+
+@pytest.mark.parametrize("user_id, cookies, project", _MY_ADMINISTERED_PROJECT_REQUESTS)
+def test_delete_own_project(
+    user_id, cookies, project, client_v2, inserted_projects_v2, db
+):
+    """Any user can delete the project they administer at /v2/me/projects/{id}"""
+    # using context manager to ensure on_startup runs
+    with client_v2 as client:
+        project_id = project["_id"]
+        url = f"/v2/me/projects/{project_id}"
+
+        assert get_db_record(db, Project, project_id) is not None
+        response = client.delete(url, cookies=cookies)
+
+        assert response.status_code == 204
+        assert get_db_record(db, Project, project_id) is None
+
+
+@pytest.mark.parametrize(
+    "user_id, cookies, project", _MY_NON_ADMINISTERED_PROJECT_REQUESTS
+)
+def test_delete_others_project_is_not_allowed(
+    user_id, cookies, project, client_v2, inserted_projects_v2, db
+):
+    """No user can delete projects they don't administer at /v2/me/projects/{id}"""
+    # using context manager to ensure on_startup runs
+    with client_v2 as client:
+        project_id = project["_id"]
+        url = f"/v2/me/projects/{project_id}"
+
+        assert get_db_record(db, Project, project_id) is not None
+        response = client.delete(url, cookies=cookies)
+
+        got = response.json()
+        expected = {"detail": "Forbidden"}
+
+        assert response.status_code == 403
+        assert got == expected
+        assert get_db_record(db, Project, project_id) is not None
 
 
 @pytest.mark.parametrize("payload", APP_TOKEN_LIST)